TST: split up test_concat.py #37243
 * created test_categorical.py
